tdf#119760 docx: table starts at left, not mid-border
related to compatibilityMode = 15; the new export default.

In Word 2013+, it appears that the table starts immediately,
and then the border is added. In Writer, the table is
considered to start mid-border. So that difference needs
to be emulated by adjusting the table margin.
The same appears to be true for nested tables.
Negative margins appear to just be ignored for nested tables,
and Word's UI does not accept a negative value.

I modified the existing table-position-15.docx slightly (in Word 2016),
just to remove spacing between the caption and the table
so that it would be easy to visually compare the table position.
One slight difference is that WORD seems to add a small "margin"
to the the paragraph shading area - don't let that throw you off.

The existing unit test enforced settings that didn't match
how Word 2013 was showing the files, despite being written to
test exactly that. I assume the author was focusing more on
OLDER versions of Word and didn't care too much about the
NEW version. A man after my own heart.

MISC:
-I removed my earlier TODO which was just plain wrong.
-the ambiguous make::any caused me untold problems,
so I made it unambiguous by moving it into a variable.

Change-Id: I0685ab74777f1def2841c7d18ff1cdaf2753f47b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91608
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>