tdf#132091 sw: fix fieldmark mess in SwCursor::SelectWordWT()
The problem is that it selects both the field instruction and the field
result; in the typical case where the instruction is empty, the
separator won't be deleted (due to CalcBreaks) but the problem is
that the cursor ends up before the separator, not after it.

Also remove some silly defensive programming.

(regression from ffb26b81e1c7ff1d64959200247bb2edd5a569da)

Change-Id: I80b6cb1790b7102828f12b680631f928734ffa7b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/92284
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>