Don't insert embedded object twice
During debugging of failure https://gerrit.libreoffice.org/c/core/+/92980,
I saw that one of objects throws calling OCommonEmbeddedObject::getEntryName
in EmbeddedObjectContainer::StoreAsChildren, because its m_bWaitSaveCompleted
is true.

The reason turns out that the object was already inserted, and thus its
OCommonEmbeddedObject::storeAsEntry was called; but not yet its
OCommonEmbeddedObject::saveCompleted. The double insertion happened in
SwXFrame::attachToRange, where the object was first passed to
InsertEmbeddedObject directly, and then was passed to InsertEmbObject,
which called the former one again in this stack:

	comphelper.dll!comphelper::EmbeddedObjectContainer::InsertEmbeddedObject(const com::sun::star::uno::Reference<com::sun::star::embed::XEmbeddedObject> & xObj, rtl::OUString & rName) Line 514	C++
	swlo.dll!SwOLEObj::SetNode(SwOLENode * pNode) Line 884	C++
	swlo.dll!SwOLENode::SwOLENode(const SwNodeIndex & rWhere, const svt::EmbeddedObjectRef & xObj, SwGrfFormatColl * pGrfColl, const SwAttrSet * pAutoAttr) Line 220	C++
	swlo.dll!SwNodes::MakeOLENode(const SwNodeIndex & rWhere, const svt::EmbeddedObjectRef & xObj, SwGrfFormatColl * pGrfColl) Line 382	C++
	swlo.dll!sw::DocumentContentOperationsManager::InsertEmbObject(const SwPaM & rRg, const svt::EmbeddedObjectRef & xObj, SfxItemSet * pFlyAttrSet) Line 2914	C++
	swlo.dll!SwXFrame::attachToRange(const com::sun::star::uno::Reference<com::sun::star::text::XTextRange> & xTextRange, const SwPaM * pCopySource) Line 3028	C++
	...

Remove the redundant call, and put a comment on that.

Change-Id: I680786a601ec454483e86c47ae07666d82c24146
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/92982
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>