tdf#60378 doc import: merged cells - don't null right border
Well, don't null them out at all, for the benefit of merged
cells, where the right border might be lost.

There doesn't seem to be any point in clearing a border because
LO sets a "merge adjacent line styles" in the borders
which does exactly what the comment indicated it was trying to
do. This option has been around since at least LO 3.5.

Change-Id: I22954e5fc31dc1c072f1ff5ccaaa05cd0f69fda4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93064
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>
Reviewed-by: Justin Luth <justin_luth@sil.org>