Add the com.apple.security.cs.disable-executable-page-protection here, too
It was added in the hardened_runtime.xcent that is used for
LibreOffice distribution for macOS outside the App Store already last
year.

Note that it is plausible that the mention of the extension manager in
the comment in hardened_runtime.xcent is a red herring, and that it is
just a coincidence that some extension manager code that runs early when
LibreOffice is starting is the first code to make use of the dynamic
machine code generation in the C++/UNO bridge.

Change-Id: I90a48d585efee27a215d77e5db45d3cda5c8fedc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93122
Tested-by: Tor Lillqvist <tml@collabora.com>
Reviewed-by: Tor Lillqvist <tml@collabora.com>