Fix row/column pre-initialization in stack blur
The changed code is used to initialize data before processing first
pixel of each row/column of the bitmap, since in order to blur it,
we need some information about assumed previous pixels. But previous
code assumed that information for the first pixel or each row and
each column is taken from first pixels of first column, not from
first pixels of respective row or respective column. This doesn't
make practical difference for the pre-extended bitmap with a solid
color of the border (see centerExtendBitmap; it's the only actual
case ATM). But in case of not pre-extended bitmap with non-solid
border color, that would result in wrong blur of first pixels.

Consider this bitmap (* is white, 0 is black):

    *00000000
    *00000000
    *00000000
    *00000000
    *00000000
    000000000
    ...

For this bitmap, existing code pre-initialized each row and each
column as if it were prepended with "******" from column 0, and so
each first pixel of each column and each row would be blurred with
white, even where it's all black (consider columns >=1, and rows
>=5).

Additionally, the existing initialization assumed that there's
always at least as many scanlines as min(radius, bmpWidth), which
might be false. See how aArrays was initialized using nLastIndexX
in stackBlurHorizontal, and then that value affected the index
passed to pReadAccess->GetScanline.

This changes the code to consistently consider first pixels of
respective rows/columns for pre-initialization data.

Change-Id: Ibe3be8750fe5347302eca782a00b36096d99820c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93118
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>