writerfilter: remove unused xCrsr
This looks like copy/paste code.
xCrsr isn't really used, and I hardly doubt that it needs
to verify that a Crsr can exist in order to ensure
that the dummy paragraph is really created.

In fact, even checking for xTextAppend.is() shouldn't
be necessary because before this it was being used
without checking. But since I myself am a copy/paste
programmer, and I see lots of examples where this
is tested in similar situations, I leave it as it is
just to be safe.

Change-Id: Ie45d4bc9d0e1cf0a0d7602b83962805165c3b85d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93413
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Michael Stahl <michael.stahl@cib.de>