More targeted silencing of -Wdeprecated-declarations
...compared to d1786724b8e8e474e1f7e39012c1f19611841dc0 "prevent warnings in
pyuno with latest python".  For one it is only the

  /* bpo-37250: kept for backwards compatibility in CPython 3.8 only */
  Py_DEPRECATED(3.8) int (*tp_print)(PyObject *, FILE *, int);

member (in /usr/include/python3.8/cpython/object.h) that causes a warning.  And
for another it is only Clang that emits a warning when initializing a deprecated
member that way, <http://lists.llvm.org/pipermail/cfe-dev/2020-May/065392.html>
"[cfe-dev] Diagnosing initialization of deprecated data member?"

Change-Id: I36625118a6bb26f5468d436da4caa82911181202
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94016
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>