tdf#95495 ooxmlexport3: less implementation-specific unit test
Although the original document completely defines the numbering
in (inherited) paragraph styles, the unit test was checking
whether the paragaph itself had the exported numbering.
That tests the current implementation where style information
is copied to the paragrah, but in a correct world that information
would stay in the styles.

This unit test uses the EXTREME example where a "chapter numbering"
is inherited by a sub-style - something LO completely does not
support, so this needs serious emulation. Thus I am keeping
the implementation-specific check, but clearly documenting
that fact.

Change-Id: I5b95646498c4d936eceb9be1841c654a5bd81275
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94401
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>