FmXListBoxCell doesn't need to implement css::awt::XListBox
FmXListBoxCell implements css::awt::XListBox but none of that functionality is
needed locally where it is used to pick a value in a cell when edited
interactively.

As an XInterface a FmXListBoxCell does fall into the
script::XEventAttacherManager attach/detach sink-hole so its very difficult to
determine if it (and any of its siblings) really need to implement the amount
of uno interfaces they actually implement, so this is a little speculative.

See https://ask.libreoffice.org/en/question/152691/how-to-populate-combo-box-within-table-control-in-form/
for an example of the ComboBox interaction case which is similar to this ListBox case.

The exotic "multiselection" option definitely isn't shown as an option in the property browser
for a listbox column in the table control.

Change-Id: I7bdc351e615c9df708a30e4396f6f352fabaed36
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94584
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>