tdf#133370 docx export: don't export unseen columnbreak
A compatibility option indicates whether a column break
is treated as a page break when there are no columns.
(Currently true only for DOCX). Don't export this column
break when compat is off, because it WILL show up as
a page-break in Word and LO when loaded.

In theory, the RTF and DOC code that converts these
into page-breaks on import could now be removed, but I
can't think of any reason why that would be of benefit.

testColumnBreak_ColumnCountIsZero in ooxmlexport4
covers the situation where docx MUST round-trip these.

Change-Id: I5c17e43d90469bdb86087c61dadf48a9440f7b96
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94799
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>