Related: tdf#133411 SetDocWin is using the previous search success state
not the new state

The order of calls probably didn't matter in the past where the use of the flag
was deferred until the Accessibility data was queried which would happen in
another event loop.

This makes it more clear that it appears that only calc actually does anything
productive here.

I think this flow-to has created more trouble that its worth and I'll remove it
but if we need to restore it, then this, I think, it the working state to
restore to.

Change-Id: Id6fbb483c081f6d5142100d70c1b29705dcb6452
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95005
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>