tdf#133453 doc import: HasSprm can return last one
The last sprm is the one that takes effect in the rare
cases of duplicate SPRMs defined, so at least have
an option to return that one as the matching value.

Unfortunately, that seems to have been ignored in the
last 20+ years of .doc support. So I added it carefully
and only enabled finding the last SPRM in the specific
case I was dealing with. However, I expect it should
~always be true, so perhaps I will followup with a
patch looking for examples of duplicate/different
SPRMs defined.

Change-Id: I5539aa3d6117380cd6d852ca165d40c7c7f3c330
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95013
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>