tdf#133271 sw textbox: handle TextRotateAngle shape property
Shape with btlr text direction is imported as TextPreRotateAngle=-270
from DOCX. Saving this to ODT turns the property name into
TextRotateAngle and its type into double.

Handle that as well to survive the ODF roundtrip of a shape+textbox
where the textbox has a btlr text direction.

(Also add a way to make multiple tests in a suite to be more independent
from each other: depending on ordering, the new test made the old test
fail. Calling ErrorRegistry::Reset() makes that go away.)

Change-Id: Iea9212f3bbb01059caf3b0f2d809e48debf52953
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95340
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins