tdf#101211 follow-up: move cursor to start of next/prev pg UNO functions
This is a follow-up patch that fixes the cursor not moving to the start
of the page when multiple pages are visible.

The cursor move behavior is; if the cursor is visible, the page number
the cursor is visible on is used as the reference. If the cursor is not
visible, the page number of the first visible page is used as the
reference.

Change-Id: Ibdfaa763bf7283e46d8c68270dfa6b5f0012ca0f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95476
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>