xmloff: ODF export: fix crash on tdf125455-2.odt
FillBitmapMode is a very funny property that is stored as 2 items
XFillBmpTileItem XATTR_FILLBMP_TILE and
XFillBmpStretchItem XATTR_FILLBMP_STRETCH
and the XPropertySet implementations for it are a bit inconsistent,
e.g. sd uses .GetItem<XFillBmpStretchItem> which searches the parent
and requires both items to be present,
but sw uses GetItemState(XATTR_FILLBMP_STRETCH, false) which does not
check the parent and one item is enough.

Let's just not export the attribute in case the property isn't defined.

(regression from 4d4404b79bf051de79f587bdafd82cc0addfc636)

Change-Id: I13c2e91c3e4eb15ebec508f679ba49614f493a11
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95501
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>