tdf#131561 writerfilter: pre-emptive unit test
Someone might be tempted to just delete any
border on a merged cell, hoping/expecting that
the adjoining cell would have a corresponding
border.

Granted, this test was hand-crafted and is not
necessarily something that Word would ever
create. At least be aware that any fix
that breaks this example is not a complete
solution.

Change-Id: I438c835b54ec8b33eb48de55f09a7cf2965f6440
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95533
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>