tdf#133859 Wizard service: disable 'Next' button if path has only base item
If the wizard dialog for extensions has only the base item in the first
path, there is no need to proceed to the next page, as there is no one.
This will be checked and if so, the 'Next' button disabled.

In libreoffice versions before 6.4, an ORoadmap class was used in the
wizard. There, if the ORoadmap data are reinitialized, the
InCompleteHyperLabel object must be destroyed first, before it will
be set to nullptr.

Change-Id: I5b4b2e6b3666b58acccace385c622f0a065fc368
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95969
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>