sd signature line: inform infobar if a cert is already selected
In case a pdf is opened for signing, the infobar has a button to start
the signing. If a signature line is already inserted and the user picked
a certificate, then let's call the infobar button's action "finish
signing", since the signature list and the certificate chooser won't
appear in this case.

(The actual behavior when you click on the button is still unchanged,
though.)

Change-Id: I50ee2455c91cdee26612a6b8239dbb6772401877
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96060
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins