macOS fix langpack installer with manually selected .app directory
For cases when there are multiple matching apps where the Languagepack
might be installed to, the "make sure to launch LO once before
installing the langpack" mechanism can fail.
Even if it is the call "tell application <selection> to activate" that
fails, the error handler assumes that the untar operation didn't work
and suggests to retry with admin privileges, modifying the .app bundle
before having it started once (AKA verified by gatekeeper), breaking
the app.
Passing the full path to the application object should fix that.

Change-Id: I47d61ad68e225fbdba191702ba85598164602e81
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96449
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>