tdf#129452 writerfilter: preserve gridSpans longer than currentRow
It is a fairly common thing for table operations to compare
items in the same column (like merged cells for example).
In order to determine column information, each row's (gridBefore
- change-id Ie305477f0e3468a4a923095d76f520d97fe99ffe -
and) merged cells need to be known.

So save that information in the cell data - and don't just throw
it away after the current row has been analyzed.  Good grief.

Change-Id: Ibfdac336bbb1f7303c7e585a85c94be37ad6f916
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/97433
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: László Németh <nemeth@numbertext.org>